% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{.wrap.GPModel}
\alias{.wrap.GPModel}
\alias{.wrap.gpb.Booster}
\alias{.unwrap.GPModel}
\alias{.unwrap.gpb.Booster}
\title{Deal with class for saving to and loading from Cache or Disk}
\usage{
.wrap.GPModel(
  obj,
  cachePath,
  preDigest,
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  outputObjects = NULL,
  ...
)

.wrap.gpb.Booster(
  obj,
  cachePath,
  preDigest,
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  outputObjects = NULL,
  ...
)

.unwrap.GPModel(
  obj,
  cachePath,
  cacheId,
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  ...
)

.unwrap.gpb.Booster(
  obj,
  cachePath,
  cacheId,
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  ...
)
}
\arguments{
\item{obj}{Any arbitrary R object.}

\item{cachePath}{A repository used for storing cached objects.
This is optional if \code{Cache} is used inside a SpaDES module.}

\item{preDigest}{The list of \code{preDigest} that comes from \code{CacheDigest} of an object}

\item{drv}{If using a database backend, \code{drv} must be an object that
inherits from \code{DBIDriver} (e.g., \code{RSQLite::SQLite}).}

\item{conn}{an optional \code{DBIConnection} object, as returned by \code{dbConnect()}.}

\item{verbose}{Numeric, -1 silent (where possible), 0 being very quiet,
1 showing more messaging, 2 being more messaging, etc.
Default is 1. Above 3 will output much more information about the internals of
Caching, which may help diagnose Caching challenges. Can set globally with an
option, e.g., \verb{options('reproducible.verbose' = 0) to reduce to minimal}}

\item{outputObjects}{Optional character vector indicating which objects to
return. This is only relevant for list, environment (or similar) objects}

\item{...}{Arguments passed to methods; default does not use anything in `...`.}

\item{cacheId}{Character string. If passed, this will override the calculated hash
of the inputs, and return the result from this \code{cacheId} in the \code{cachePath}.
Setting this is equivalent to manually saving the output of this function, i.e.,
the object will be on disk, and will be recovered in subsequent
This may help in some particularly finicky situations
where \code{Cache} is not correctly detecting unchanged inputs. This will guarantee
the object will be identical each time; this may be useful in operational code.}
}
\value{
Returns an object that can be saved to disk e.g., via `saveRDS`.
}
\description{
Methods that prepare `GPModel` and `gpb.Booster` objcts for
saving to disk (or RAM) via e.g., `saveRDS`, which need wrapping and unwrapping.
}
