% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{.wrap}
\alias{.wrap}
\alias{.wrap.GPModel}
\alias{.wrap.gpb.Booster}
\alias{.unwrap}
\alias{.unwrap.GPModel}
\alias{.unwrap.gpb.Booster}
\title{Deal with class for saving to and loading from Cache or Disk}
\usage{
.wrap(
  obj,
  cachePath,
  preDigest,
  drv = reproducible:::getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  outputObjects = NULL,
  ...
)

\method{.wrap}{GPModel}(
  obj,
  cachePath,
  preDigest,
  drv = reproducible:::getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  outputObjects = NULL,
  ...
)

\method{.wrap}{gpb.Booster}(
  obj,
  cachePath,
  preDigest,
  drv = reproducible:::getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  outputObjects = NULL,
  ...
)

.unwrap(
  obj,
  cachePath,
  cacheId,
  drv = reproducible:::getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  ...
)

\method{.unwrap}{GPModel}(
  obj,
  cachePath,
  cacheId,
  drv = reproducible:::getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  ...
)

\method{.unwrap}{gpb.Booster}(
  obj,
  cachePath,
  cacheId,
  drv = reproducible:::getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  ...
)
}
\arguments{
\item{obj}{Any arbitrary R object.}

\item{cachePath}{A repository used for storing cached objects.
This is optional if \code{Cache} is used inside a SpaDES module.}

\item{preDigest}{The list of \code{preDigest} that comes from \code{CacheDigest} of an object}

\item{drv}{If using a database backend, \code{drv} must be an object that
inherits from \code{DBIDriver} (e.g., \code{RSQLite::SQLite}).}

\item{conn}{an optional \code{DBIConnection} object, as returned by \code{dbConnect()}.}

\item{verbose}{Numeric, -1 silent (where possible), 0 being very quiet,
1 showing more messaging, 2 being more messaging, etc.
Default is 1. Above 3 will output much more information about the internals of
Caching, which may help diagnose Caching challenges. Can set globally with an
option, e.g., \verb{options('reproducible.verbose' = 0) to reduce to minimal}}

\item{outputObjects}{Optional character vector indicating which objects to
return. This is only relevant for list, environment (or similar) objects}

\item{...}{Arguments passed to methods; default does not use anything in \code{...}.}

\item{cacheId}{Used strictly for messaging. This should be the cacheId of the object being recovered.}
}
\description{
See \code{\link[reproducible:dotWrap]{reproducible::.wrap()}}
}
